Reformulating VillageStats & Tasks with ThreadLoop

+++ Application/src/Application/Game/Ingame/RobotFunctions.hpp
@@ -6,7 +6,7 @@ namespace Application
 {
     const int FUNCTION_QTY = 5;
 
-    enum class RobotFunction : short
+    enum class RobotFunction : uint8_t
     {
         PROTECTION = 0,
         HUNT,
